use new extension account utils, discontinue `Extensions` class
