+++ library/react-connect-kit/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@w3ux/react-connect-kit-source",
   "license": "GPL-3.0-only",
-  "version": "3.2.2",
+  "version": "3.2.3",
   "type": "module",
   "scripts": {
     "clear": "rm -rf node_modules dist tsconfig.tsbuildinfo",
@@ -9,13 +9,13 @@
   },
   "dependencies": {
     "@w3ux/extension-assets": "^2.2.0",
-    "@w3ux/hooks": "^2.0.4",
-    "@w3ux/observables-connect": "0.9.1",
+    "@w3ux/hooks": "^2.1.0",
+    "@w3ux/observables-connect": "^0.9.3",
     "@w3ux/utils": "^2.0.8"
   },
   "devDependencies": {
     "@types/react": "^19.0.12",
-    "@w3ux/types": "^2.1.1",
+    "@w3ux/types": "^2.1.2",
     "gulp": "^5.0.0",
     "gulp-sourcemaps": "^3.0.0",
     "gulp-strip-comments": "^2.6.0",
