+++ library/react-connect-kit/src/ExtensionAccountsProvider/index.tsx
@@ -2,6 +2,18 @@
 SPDX-License-Identifier: GPL-3.0-only */
 
 import { createSafeContext, useEffectIgnoreInitial } from '@w3ux/hooks'
+import {
+  connectedExtensions,
+  enableExtensions,
+  extensionsWithError,
+  formatEnabledExtensions,
+  getAccountsFromExtensions,
+  getExtensionsById,
+} from '@w3ux/observables-connect/accounts'
+import {
+  addExtensionToLocal,
+  removeExtensionFromLocal,
+} from '@w3ux/observables-connect/extensions/local'
 import type {
   ExtensionAccount,
   ExtensionInterface,
@@ -16,7 +28,6 @@ import {
 } from '@w3ux/utils'
 import { useEffect, useRef, useState } from 'react'
 import { useExtensions } from '../ExtensionsProvider'
-import { Extensions } from './Extensions'
 import type {
   ExtensionAccountsContextInterface,
   ExtensionAccountsProviderProps,
@@ -50,33 +61,33 @@ export const ExtensionAccountsProvider = ({
     extensionHasFeature,
   } = useExtensions()
 
-  // Store connected extension accounts.
+  // Store connected extension accounts
   const [extensionAccounts, setExtensionAccounts] = useState<ImportedAccount[]>(
     []
   )
   const extensionAccountsRef = useRef(extensionAccounts)
 
-  // Store whether extension accounts have been synced.
+  // Store whether extension accounts have been synced
   const [extensionAccountsSynced, setExtensionAccountsSynced] =
     useState<Sync>('unsynced')
 
-  // Store extensions whose account subscriptions have been initialised.
+  // Store extensions whose account subscriptions have been initialised
   const [extensionsInitialised, setExtensionsInitialised] = useState<string[]>(
     []
   )
   const extensionsInitialisedRef = useRef(extensionsInitialised)
 
-  // Store unsubscribe handlers for connected extensions.
+  // Store unsubscribe handlers for connected extensions
   const unsubs = useRef<Record<string, VoidFn>>({})
 
-  // Helper for setting active account. Ignores if not a valid function.
+  // Helper for setting active account. Ignores if not a valid function
   const maybeSetActiveAccount = (address: string) => {
     if (typeof setActiveAccount === 'function') {
       setActiveAccount(address ?? null)
     }
   }
 
-  // Helper for calling extension enabled callback. Ignores if not a valid function.
+  // Helper for calling extension enabled callback. Ignores if not a valid function
   const maybeOnExtensionEnabled = (id: string) => {
     if (typeof onExtensionEnabled === 'function') {
       onExtensionEnabled(id)
@@ -92,42 +103,40 @@ export const ExtensionAccountsProvider = ({
   // Connects to extensions that already have been connected to and stored in localStorage. Loop
   // through extensions and connect to accounts. If `activeAccount` exists locally, we wait until
   // all extensions are looped before connecting to it; there is no guarantee it still exists - must
-  // explicitly find it.
+  // explicitly find it
   const connectActiveExtensions = async () => {
     const extensionIds = Object.keys(extensionsStatus)
     if (!extensionIds.length) {
       return
     }
 
-    // Iterate previously connected extensions and retreive valid `enable` functions.
+    // Iterate previously connected extensions and retreive valid `enable` functions
     // ------------------------------------------------------------------------------
-    const rawExtensions = Extensions.getFromIds(extensionIds)
+    const rawExtensions = getExtensionsById(extensionIds)
 
-    // Attempt to connect to extensions via `enable` and format the results.
-    const enableResults = Extensions.formatEnabled(
+    // Attempt to connect to extensions via `enable` and format the results
+    const enableResults = formatEnabledExtensions(
       rawExtensions,
-      await Extensions.enable(rawExtensions, dappName)
+      await enableExtensions(rawExtensions, dappName)
     )
 
-    // Retrieve the resulting connected extensions only.
-    const connectedExtensions = Extensions.connected(enableResults)
+    // Retrieve the resulting connected extensions only
+    const connected = connectedExtensions(enableResults)
 
-    // Retrieve  extensions that failed to connect.
-    const extensionsWithError = Extensions.withError(enableResults)
+    // Retrieve extensions that failed to connect
+    const withError = extensionsWithError(enableResults)
 
     // Add connected extensions to local storage.
-    Array.from(connectedExtensions.keys()).forEach((id) =>
-      Extensions.addToLocal(id)
-    )
+    Array.from(connected.keys()).forEach((id) => addExtensionToLocal(id))
 
-    // Initial fetch of extension accounts to populate accounts & extensions state.
+    // Initial fetch of extension accounts to populate accounts & extensions state
     // ----------------------------------------------------------------------------
 
     // Get full list of imported accounts.
-    const initialAccounts = await Extensions.getAllAccounts(connectedExtensions)
+    const initialAccounts = await getAccountsFromExtensions(connected)
 
     // Get the active account if found in initial accounts. Format initial account addresses to the
-    // correct ss58 format before finding.
+    // correct ss58 format before finding
     const activeAccountInInitial = initialAccounts
       .map((acc) => ({
         ...acc,
@@ -135,24 +144,24 @@ export const ExtensionAccountsProvider = ({
       }))
       .find(({ address }) => address === getActiveAccountLocal(network, ss58))
 
-    // Perform all initial state updates.
+    // Perform all initial state updates
     // ----------------------------------
 
-    Array.from(extensionsWithError.entries()).forEach(([id, state]) => {
+    Array.from(withError.entries()).forEach(([id, state]) => {
       handleExtensionError(id, state.error)
     })
 
-    Array.from(connectedExtensions.keys()).forEach((id) => {
+    Array.from(connected.keys()).forEach((id) => {
       setExtensionStatus(id, 'connected')
       updateInitialisedExtensions(id)
     })
 
     updateExtensionAccounts({ add: initialAccounts, remove: [] })
 
-    // Initiate account subscriptions for connected extensions.
+    // Initiate account subscriptions for connected extensions
     // --------------------------------------------------------
 
-    // Handler function for each extension accounts subscription.
+    // Handler function for each extension accounts subscription
     const handleAccounts = (
       extensionId: string,
       accounts: ExtensionAccount[],
@@ -170,25 +179,23 @@ export const ExtensionAccountsProvider = ({
         ss58
       )
 
-      // Update added and removed accounts.
+      // Update added and removed accounts
       updateExtensionAccounts({ add: newAccounts, remove: accountsToRemove })
     }
 
-    // Try to subscribe to accounts for each connected extension.
-    for (const [id, { extension }] of Array.from(
-      connectedExtensions.entries()
-    )) {
+    // Try to subscribe to accounts for each connected extension
+    for (const [id, { extension }] of Array.from(connected.entries())) {
       // If enabled, subscribe to accounts.
       if (extensionHasFeature(id, 'subscribeAccounts')) {
         const unsub = extension.accounts.subscribe((accounts) => {
           handleAccounts(id, accounts || [], extension.signer)
         })
-        // Add unsub to context ref.
+        // Add unsub to context ref
         addToUnsubscribe(id, unsub)
       }
     }
 
-    // Connect to active account if found in initial accounts.
+    // Connect to active account if found in initial accounts
     if (activeAccountInInitial) {
       connectActiveExtensionAccount(activeAccountInInitial, connectToAccount)
     }
@@ -197,7 +204,7 @@ export const ExtensionAccountsProvider = ({
   // connectExtensionAccounts
   //
   // Similar to the above but only connects to a single extension. This is invoked by the user by
-  // clicking on an extension. If activeAccount is not found here, it is simply ignored.
+  // clicking on an extension. If activeAccount is not found here, it is simply ignored
   const connectExtensionAccounts = async (id: string): Promise<boolean> => {
     const extensionIds = Object.keys(extensionsStatus)
     const exists = extensionIds.find((key) => key === id) || undefined
@@ -208,15 +215,15 @@ export const ExtensionAccountsProvider = ({
       )
     } else {
       try {
-        // Attempt to get extension `enable` property.
+        // Attempt to get extension `enable` property
         const { enable } = window.injectedWeb3[id]
 
         // Summons extension popup.
         const extension: ExtensionInterface = await enable(dappName)
 
-        // Continue if `enable` succeeded, and if the current network is supported.
+        // Continue if `enable` succeeded, and if the current network is supported
         if (extension !== undefined) {
-          // Handler for new accounts.
+          // Handler for new accounts
           const handleAccounts = (accounts: ExtensionAccount[]) => {
             const {
               newAccounts,
@@ -229,7 +236,7 @@ export const ExtensionAccountsProvider = ({
               network,
               ss58
             )
-            // Set active account for network if not yet set.
+            // Set active account for network if not yet set
             if (!activeAccount) {
               const activeExtensionAccount = getActiveExtensionAccount(
                 network,
@@ -247,23 +254,23 @@ export const ExtensionAccountsProvider = ({
               }
             }
 
-            // Update extension accounts state.
+            // Update extension accounts state
             updateExtensionAccounts({
               add: newAccounts,
               remove: accountsToRemove,
             })
 
-            // Update initialised extensions.
+            // Update initialised extensions
             updateInitialisedExtensions(id)
           }
 
-          // Call optional `onExtensionEnabled` callback.
-          Extensions.addToLocal(id)
+          // Call optional `onExtensionEnabled` callback
+          addExtensionToLocal(id)
 
           maybeOnExtensionEnabled(id)
           setExtensionStatus(id, 'connected')
 
-          // If account subscriptions are not supported, simply get the account(s) from the extension. Otherwise, subscribe to accounts.
+          // If account subscriptions are not supported, simply get the account(s) from the extension. Otherwise, subscribe to accounts
           if (!extensionHasFeature(id, 'subscribeAccounts')) {
             const accounts = await extension.accounts.get()
             handleAccounts(accounts)
@@ -282,26 +289,26 @@ export const ExtensionAccountsProvider = ({
     return false
   }
 
-  // Handle errors when communiating with extensions.
+  // Handle errors when communiating with extensions
   const handleExtensionError = (id: string, err: string) => {
-    // if not general error (maybe enabled but no accounts trust app).
+    // if not general error (maybe enabled but no accounts trust app)
     if (err.startsWith('Error')) {
-      // remove extension from local `active_extensions`.
-      Extensions.removeFromLocal(id)
+      // remove extension from local `active_extensions`
+      removeExtensionFromLocal(id)
 
       // extension not found (does not exist).
       if (err.substring(0, 17) === 'NotInstalledError') {
         removeExtensionStatus(id)
       } else {
-        // declare extension as no imported accounts authenticated.
+        // declare extension as no imported accounts authenticated
         setExtensionStatus(id, 'not_authenticated')
       }
     }
-    // mark extension as initialised.
+    // mark extension as initialised
     updateInitialisedExtensions(id)
   }
 
-  // Update initialised extensions.
+  // Update initialised extensions
   const updateInitialisedExtensions = (id: string) => {
     if (!extensionsInitialisedRef.current.includes(id)) {
       setStateWithRef(
@@ -312,7 +319,7 @@ export const ExtensionAccountsProvider = ({
     }
   }
 
-  // Add an extension account to context state.
+  // Add an extension account to context state
   const updateExtensionAccounts = ({
     add,
     remove,
@@ -320,16 +327,16 @@ export const ExtensionAccountsProvider = ({
     add: ExtensionAccount[]
     remove: ExtensionAccount[]
   }) => {
-    // Add new accounts and remove any removed accounts.
+    // Add new accounts and remove any removed accounts
     const newAccounts = [...extensionAccountsRef.current]
       .concat(add)
       .filter((a) => remove.find((s) => s.address === a.address) === undefined)
 
     if (remove.length) {
-      // Unsubscribe from removed accounts.
+      // Unsubscribe from removed accounts
       unsubAccounts(remove)
 
-      // Remove active account if it is being forgotten.
+      // Remove active account if it is being forgotten
       if (
         activeAccount &&
         remove.find(({ address }) => address === activeAccount) !== undefined
@@ -341,14 +348,14 @@ export const ExtensionAccountsProvider = ({
     setStateWithRef(newAccounts, setExtensionAccounts, extensionAccountsRef)
   }
 
-  // Add an extension id to unsubscribe state.
+  // Add an extension id to unsubscribe state
   const addToUnsubscribe = (id: string, unsub: VoidFn) => {
     unsubs.current[id] = unsub
   }
 
-  // Handle unsubscribing of an removed extension accounts.
+  // Handle unsubscribing of an removed extension accounts
   const unsubAccounts = (accounts: ImportedAccount[]) => {
-    // Unsubscribe and remove unsub from context ref.
+    // Unsubscribe and remove unsub from context ref
     if (accounts.length) {
       for (const { address } of accounts) {
         if (extensionAccountsRef.current.find((a) => a.address === address)) {
@@ -362,7 +369,7 @@ export const ExtensionAccountsProvider = ({
     }
   }
 
-  // Unsubscrbe all account subscriptions.
+  // Unsubscrbe all account subscriptions
   const unsubscribe = () => {
     Object.values(unsubs.current).forEach((unsub) => {
       unsub()
@@ -370,14 +377,14 @@ export const ExtensionAccountsProvider = ({
   }
 
   const handleSyncExtensionAccounts = async () => {
-    // Wait for injectedWeb3 check to finish before starting account import process.
+    // Wait for injectedWeb3 check to finish before starting account import process
     if (!checkingInjectedWeb3 && extensionAccountsSynced === 'unsynced') {
       // Unsubscribe from all accounts and reset state
       unsubscribe()
       setStateWithRef([], setExtensionAccounts, extensionAccountsRef)
       setStateWithRef([], setExtensionsInitialised, extensionsInitialisedRef)
       // If extensions have been fetched, get accounts if extensions exist and local extensions
-      // exist (previously connected).
+      // exist (previously connected)
       if (Object.keys(extensionsStatus).length) {
         // get active extensions
         const localExtensions = localStorageOrDefault(
@@ -391,12 +398,12 @@ export const ExtensionAccountsProvider = ({
         }
       }
 
-      // Syncing is complete. Also covers case where no extensions were found.
+      // Syncing is complete. Also covers case where no extensions were found
       setExtensionAccountsSynced('synced')
     }
   }
 
-  // Get extension accounts based on the provided ss58 prefix.
+  // Get extension accounts based on the provided ss58 prefix
   const getExtensionAccounts = (ss58Prefix: number): ImportedAccount[] =>
     extensionAccounts
       .map((account) => {
@@ -409,10 +416,10 @@ export const ExtensionAccountsProvider = ({
           address: formattedAddress,
         }
       })
-      // Remove null entries resulting from invalid formatted addresses.
+      // Remove null entries resulting from invalid formatted addresses
       .filter((account) => account !== null)
 
-  // Re-sync extensions accounts on `unsynced`.
+  // Re-sync extensions accounts on `unsynced`
   useEffect(() => {
     handleSyncExtensionAccounts()
 
@@ -420,7 +427,7 @@ export const ExtensionAccountsProvider = ({
   }, [extensionsStatus, checkingInjectedWeb3, extensionAccountsSynced])
 
   // Once initialised extensions equal total extensions present in `injectedWeb3`, mark extensions
-  // as fetched.
+  // as fetched
   useEffectIgnoreInitial(() => {
     if (
       !checkingInjectedWeb3 &&
